+++ Lesson_10/Units/ManClass.java
@@ -4,7 +4,7 @@ public abstract class ManClass extends BaseHero {
     protected int poisons;
     protected int dexterity;
 
-    public ManClass(int hp, int speed, int damage, int poisons, int dexterity, String name) {
+    public ManClass(int hp, int speed, float damage, int poisons, int dexterity, String name) {
         super(hp, speed, damage, name);
         this.poisons = poisons;
         this.dexterity = dexterity;
